+++ src/index.ts
@@ -1,6 +1,12 @@
 import Epub from './epub.js';
-import type { Options, Metadata, Section } from './types.js';
+import type {
+  CoverType,
+  Metadata,
+  Options,
+  Resource,
+  Section,
+} from './types.js';
 
 /* eslint-disable import/no-unused-modules */
 export default Epub;
-export type { Options, Metadata, Section };
+export type { CoverType, Metadata, Options, Resource, Section };