+++ src/pug.ts
@@ -5,11 +5,25 @@ import type { Data, Section } from './types.js';
 const { url } = import.meta;
 const getPath = (path: string) => new URL(path, url).pathname;
 
-const pugContainer = pug.compileFile(getPath('../templates/container.pug'));
-const pugContents = pug.compileFile(getPath('../templates/contents.pug'));
-const pugCover = pug.compileFile(getPath('../templates/cover.pug'));
-const pugOpf = pug.compileFile(getPath('../templates/opf.pug'));
-const pugSection = pug.compileFile(getPath('../templates/section.pug'));
+const options = {
+  doctype: 'xml',
+  pretty: true,
+};
+
+const pugContainer = pug.compileFile(
+  getPath('../templates/container.pug'),
+  options,
+);
+const pugContents = pug.compileFile(
+  getPath('../templates/contents.pug'),
+  options,
+);
+const pugCover = pug.compileFile(getPath('../templates/cover.pug'), options);
+const pugOpf = pug.compileFile(getPath('../templates/opf.pug'), options);
+const pugSection = pug.compileFile(
+  getPath('../templates/section.pug'),
+  options,
+);
 
 // Provide the contents of the container XML file.
 const getContainer = () => pugContainer();
@@ -23,13 +37,14 @@ const processTextContent = (content: string) =>
 // Provide the contents of the cover HTML enclosure.
 const getCover = (data: Data) => {
   const { coverType } = data.options;
+  const { cover } = data;
 
-  let coverText = data.metadata.cover;
-  if (coverType === 'text') {
-    coverText = processTextContent(data.cover.text);
+  if (coverType === 'text' && typeof cover === 'string') {
+    const coverText = processTextContent(cover);
+    return pugCover({ coverText, data });
   }
 
-  return pugCover({ coverText, data });
+  return pugCover({ data });
 };
 
 // Provide the contents of the OPF (spine) file.