+++ src/types.ts
@@ -10,9 +10,21 @@ type Section = {
   isFrontMatter: boolean;
 }>;
 
+type Resource = {
+  name: string;
+  data: Buffer;
+} & Partial<{
+  // This is base of file, and cannot be overriden
+  base: string;
+  // Any properties to apply to this resource, i.e. cover-image
+  properties: string;
+  // This is the mime type, and can be overriden by user
+  type: string;
+}>;
+
 type BaseMetadata = {
   author: string;
-  cover: string;
+  cover: string | Resource;
   id: number | string;
   title: string;
 };
@@ -34,26 +46,17 @@ type Metadata = BaseMetadata &
     tags: string[];
   }>;
 
-type Image = {
-  base: string;
-  originalFilename: string;
-  type: string;
-};
-
 type Options = Partial<{
   coverType: CoverType;
   showContents: boolean;
 }>;
 
 type Data = {
-  cover: {
-    image: Image;
-    text: string;
-  };
+  cover: string | Required<Resource>;
   css: string;
-  images: Image[];
   metadata: Required<Metadata>;
   options: Required<Options>;
+  resources: Required<Resource>[];
   sections: Required<Section>[];
 };
 
@@ -62,8 +65,8 @@ type Document = {
   sections: Section[];
 } & Partial<{
   css: string;
-  images: string[];
   options: Options;
+  resources: Resource[];
 }>;
 
-export type { Data, Document, Metadata, Options, Section };
+export type { CoverType, Data, Document, Metadata, Options, Resource, Section };